🚚 rename [#28] 파일 이름과 경로가 중복될 때 파일 이름을 `index`로 변경
